requirements.txt
asgiref==3.4.1
Django==3.2.8
powerline-status==2.7
PyMySQL==1.0.2
pytz==2021.3
sqlparse==0.4.2
# 已通过系统包安装，避免重复安装和编译
# cryptography==41.0.5
# boto3==1.34.72
django-cors-headers==4.3.0
djangorestframework==3.14.0
django-storages==1.14.2
PyJWT==2.8.0
requests==2.31.0
